% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module 2.R
\name{consulter_taches}
\alias{consulter_taches}
\title{Consulter les tâches assignées par la chercheuse de l'aventure 2}
\usage{
consulter_taches(section)
}
\arguments{
\item{section}{Un caractère spécifiant la section dont on veut obtenir la tâche. Options disponibles :
- `"histogramme"` : Analyse de la distribution des longueurs de nageoires.
- `"statistiques_descriptives"` : Calcul des statistiques descriptives des longueurs de nageoires.}
}
\value{
Une chaîne de caractères contenant la description de la tâche assignée.
}
\description{
Cette fonction permet d'obtenir les tâches laissées par la chercheuse en fonction de la section d'analyse.
}
\examples{
consulter_taches("histogramme")
consulter_taches("statistiques_descriptives")

}
