% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_tidyverse_style.R
\name{eval_tidyverse_style}
\alias{eval_tidyverse_style}
\title{Évaluer le style tidyverse d'un fichier R ou QMD}
\usage{
eval_tidyverse_style(file_path)
}
\arguments{
\item{file_path}{Chemin vers le fichier à analyser (format `.R` ou `.qmd`).}
}
\value{
Une liste contenant :
  - `scores` : Une liste avec les notes attribuées à chaque critère.
  - `total` : Le score total sur 20.
}
\description{
Cette fonction analyse un fichier `.R` ou `.qmd` et attribue un score basé sur les critères
du guide de style tidyverse (chapitres 1 et 2).
}
\examples{
\dontrun{
result <- eval_tidyverse_style("mon_script.R")
print(result)
}

}
