% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module 2.R
\name{verifier_valeur_aberrante}
\alias{verifier_valeur_aberrante}
\title{Vérifier si une valeur est une aberration dans la base de données penguins_mission}
\usage{
verifier_valeur_aberrante(ligne)
}
\arguments{
\item{ligne}{Un entier représentant le numéro de ligne à vérifier.}
}
\value{
Un message indiquant si la valeur correspond bien à une aberration ou si une autre vérification est nécessaire.
}
\description{
Cette fonction permet de vérifier si une ligne donnée correspond à une valeur aberrante détectée dans les données des manchots.
}
\examples{
verifier_valeur_aberrante(10) # Indique que la valeur n'est pas une aberration

}
